+++ packages/web-components/@sfx/sayt/stories/index.ts
@@ -1,17 +1,17 @@
 import { storiesOf } from '@storybook/html';
-import { withKnobs, text, boolean } from '@storybook/addon-knobs';
-import { PRODUCTS_EVENT } from '@sfx/products';
-import { AUTOCOMPLETE_RECEIVED_RESULTS_EVENT } from '@sfx/autocomplete';
+import { withKnobs, text, boolean, number } from '@storybook/addon-knobs';
+import { PRODUCTS_RESPONSE_EVENT, PRODUCTS_REQUEST_EVENT } from '@sfx/products';
+import { AUTOCOMPLETE_RECEIVED_RESULTS_EVENT, AUTOCOMPLETE_REQUEST_RESULTS } from '@sfx/autocomplete';
+import { SAYT_EVENT } from '../src/events';
+import '../src';
 import {
   getDisplayCode,
   productsResultsEvent,
   autocompleteReceivedResultsEvent,
   autocompleteResults,
   getProductsReceivedEvent,
-  hidePrompt
+  hidePrompt,
 } from '../../../../../.storybook/common';
-import { SAYT_EVENT } from '../src/events';
-import '../src';
 
 const saytNotesMarkdownIntro = ` # SF-X SAYT Component
 
@@ -32,6 +32,7 @@ function getSayt(searchbox: string = ''): string {
   const showCloseButton = boolean('Show Close button', true) ? 'showclosebutton' : '';
   const hideAutocomplete = boolean('Hide Autocomplete', false) ? 'hideAutocomplete' : '';
   const hideProducts = boolean('Hide Products', false) ? 'hideProducts' : '';
+  const minSearchLength = number('Min search length', 5);
 
   return (
     '<sfx-sayt\n' +
@@ -40,6 +41,7 @@ function getSayt(searchbox: string = ''): string {
     (showCloseButton ? `  ${showCloseButton}\n` : '') +
     (hideAutocomplete ? `  ${hideAutocomplete}\n` : '') +
     (hideProducts ? `  ${hideProducts}\n` : '') +
+    `  minsearchlength="${minSearchLength}"\n` +
     '></sfx-sayt>'
   );
 }
@@ -56,15 +58,18 @@ const saytShow = {
 
 const autocompleteDataReceivedEvent = new CustomEvent(AUTOCOMPLETE_RECEIVED_RESULTS_EVENT, {
   detail: {
-    results: autocompleteResults
-  }
+    results: autocompleteResults,
+  },
+  bubbles: true,
 });
 
 function generateBaseData() {
   setTimeout(() => {
-    window.dispatchEvent(autocompleteDataReceivedEvent)}, 0)
+    window.dispatchEvent(autocompleteDataReceivedEvent);
+  }, 0)
   setTimeout(() => {
-    window.dispatchEvent(getProductsReceivedEvent())}, 0)
+    window.dispatchEvent(getProductsReceivedEvent());
+  }, 0)
 }
 
 storiesOf('Components|SAYT', module)
@@ -86,16 +91,16 @@ storiesOf('Components|SAYT', module)
     ${getDisplayCode(sayt)}
     </div>`;
   }, {
-    notes: {
-      markdown: `
+      notes: {
+        markdown: `
         ${saytNotesMarkdownIntro}
 
           ### The SF-X SAYT component populated with autocomplete and products data.
           * The SF-X SAYT component is designed to render different data in response to events, however, this story generates the data on page load. This is done in order to see the visual component immediately.
           * For stories that demonstrate the component's functionality, visit three stories listed below this one, under "SAYT".
       `
-    }
-  })
+      }
+    })
   .add(
     'Rendering based on events and attributes',
     () => {
@@ -118,7 +123,7 @@ storiesOf('Components|SAYT', module)
       <div class="display-code">
       ${getDisplayCode(sayt)}
       </div>`
-      ;
+        ;
     },
     {
       customEvents: [productsResultsEvent, autocompleteReceivedResultsEvent, saytHide, saytShow],
@@ -134,9 +139,9 @@ storiesOf('Components|SAYT', module)
               3. See the SF-X SAYT component update with the autocomplete data.
 
 
-          ### The component will render the SF-X Products component with the payload of the \`${PRODUCTS_EVENT}\`, in response to the event.
+          ### The component will render the SF-X Products component with the payload of the \`${PRODUCTS_RESPONSE_EVENT}\`, in response to the event.
             * To emit the event in this story:
-              1. Visit the **Custom Events** tab and locate the \`${PRODUCTS_EVENT}\` event.
+              1. Visit the **Custom Events** tab and locate the \`${PRODUCTS_RESPONSE_EVENT}\` event.
               2. Click "emit".
               3. See the SF-X SAYT component update with the products data.
 
@@ -170,6 +175,8 @@ storiesOf('Components|SAYT', module)
               * \`hideAutocomplete\`
               * \`hideProducts\`
               * \`closeText\`
+              * \`minSearchLength\`
+                * **Note**: to observe the impact of modifying this attribute, it is necessary to add an event listener to listen for either the \`${AUTOCOMPLETE_REQUEST_RESULTS} \`or \`${PRODUCTS_REQUEST_EVENT}\` events. These events will trigger when the number of characters typed into the input box is equal to or greater than the number defined in the attribute.
             * To modify these attributes:
               1. Visit the **Knobs** tab and click any of the "Hide Autocomplete", "Hide Products", and "Show Close button'", and/or update the text contained within the "Close link text" field.
               2. Emit the appropriate events