+++ packages/web-components/@sfx/search-box/src/index.ts
@@ -1,1 +1,2 @@
 export { default as SearchBox } from './search-box';
+export { SEARCHBOX_EVENT } from './events';